fixed all warnings

+++ src/db/foodbatch_db.c
@@ -511,7 +511,7 @@ int foodbatch_db_get_all_format(database *db, char *buffer, size_t buffer_size) 
     }
 
     sqlite3_finalize(stmt);
-    return written;
+    return (int)written;
 }
 
 char *foodbatch_db_get_all_format_old(database *db) {
